Add files via upload
Adds missing gold walltulas in order to allow for 100% gold walltula completion and fixes some errors like the waterfall flowing upwards instead of downwards in the first area of the game. Also reverts item and area names in the pause screen to vanilla but the names remain unique in game when the title area cards show up upon entering them.